+++ src/components/policy/filters/actorTypeFilter.js
@@ -99,10 +99,30 @@ export default function ActorTypeFilter({ data, status, setstatus, filterValue }
               options={data}
               defaultValue={filterValue.value}
               getOptionLabel={(option) => option.name}
-              renderInput={(params) => <TextField {...params} label="Actor type" placeholder="Actor type" />}
+              renderInput={(params) => (
+                <TextField
+                  {...params}
+                  label={
+                    <Trans
+                      i18nKey="policies.filters.actor_type"
+                      values={{
+                        name: ''
+                      }}
+                    />
+                  }
+                  placeholder={
+                    <Trans
+                      i18nKey="policies.filters.actor_type"
+                      values={{
+                        name: ''
+                      }}
+                    />
+                  }
+                />
+              )}
               onChange={(event, value) => filterValue.set(value)}
               isOptionEqualToValue={(option, value) => option.iri === value.iri}
-              sx={{ width: '100%', marginTop: '10px' }}
+              sx={{ width: '100%', marginTop: '2%' }}
             />
           </Grow>
         </Grid>